+++ odoo/local-src/specific_helpdesk/models/helpdesk.py
@@ -13,7 +13,7 @@ class HelpdeskTicket(models.Model):
         default='/'
     )
     ticket_type_id = fields.Many2one(
-            default=lambda self: self.env.ref('helpdesk.type_incident')
+        default=lambda self: self.env.ref('helpdesk.type_incident')
     )
     helpdesk_ticket_reason_id = fields.Many2one(
         comodel_name='helpdesk.ticket.reason', string='Reason',
@@ -22,6 +22,13 @@ class HelpdeskTicket(models.Model):
     stock_picking_id = fields.Many2one(
         comodel_name='stock.picking',
         string='Stock picking',
+    )
+    stock_move_id = fields.Many2one(
+        comodel_name='stock.move',
+        string='Stock move',
+    )
+    lots = fields.Many2one(
+        related='stock_move_id.quant_ids.lot_id'
     )
     sale_order_id = fields.Many2one(
         comodel_name='sale.order',
@@ -58,8 +65,31 @@ class HelpdeskTicket(models.Model):
             )
         if partners_to_add:
             ticket.message_subscribe(partner_ids=partners_to_add)
-
         return ticket
+
+    @api.model
+    def new_one(self, r):
+        """Return the action for the wizard to create a new ticket."""
+        view = self.env.ref(
+                'specific_helpdesk.create_helpdesk_ticket_view_form')
+        return {
+            'type': 'ir.actions.act_window',
+            'res_model': 'create.helpdesk.ticket',
+            'res_id': r.id,
+            'view_id': view.id,
+            'view_type': 'form',
+            'view_mode': 'form',
+            'target': 'new',
+            }
+
+    @api.model
+    def show_existing(self, domain):
+        """Show the helpdesk tickets for a specific domain."""
+        action_data = self.env.ref(
+            'helpdesk.helpdesk_ticket_action_main_tree'
+        ).read()[0]
+        action_data['domain'] = domain
+        return action_data
 
 
 class HelpdeskTicketReason(models.Model):