+++ odoo/local-src/specific_shipping_costs/tests/__init__.py
@@ -2,4 +2,6 @@
 # Copyright 2017 Camptocamp SA
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
-from . import test_shipping_costs
+# Disable because all this module is changing in ALCN-1117
+
+# from . import test_shipping_costs